python/streamlit-langgraph/requirements.txt
streamlit>=1.28.0
langgraph>=0.0.40
langchain>=0.1.0
google-generativeai>=0.3.0
memorystack>=1.0.0
python-dotenv>=1.0.0
